don't spam chat if harvesting
closes #54
closes #52

Checks if the event's item (in player hand usually) maps to a plant
material, and if that material is the same as the block being hit
(broken in case of harvest), then cancels further interaction including
growth simulation.

@soerxpso luckily we already had a item-to-block map internally, so we
don't seem to need to track players. This also removes code with unsafe
cast and IndexOutOfBounds: avoid casting, or if needed do a runtime
`instanceof` check or similar; and mind that lists can normally be empty
so you need to check before accessing indices.